no crc errors in status; closing the gap?
